ingestion/requirements.txt
# ingestion/requirements.txt

# Same as backend requirements
openai==1.10.0  # Optional: only needed if using OpenAI embeddings
pinecone-client>=3.0.0  # Supports both old and new SDK
sentence-transformers==2.3.1  # For local embeddings (recommended)
langchain==0.1.4
langchain-community==0.0.16
unstructured==0.11.8
pypdf==4.0.0
python-docx==1.1.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0